pyproject.toml
[project]
name = "altoro-playwright-automation"
version = "1.0.0"
description = "Production-grade Playwright automation for AltoroMutual banking demo"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.2",
    "Pillow>=10.1.0",
    "python-dateutil>=2.8.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer[all]>=0.9.0",
    "python-dotenv>=1.0.0",
    "backoff>=2.2.1",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.10.0",
    "mypy>=1.6.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
    "S",  # flake8-bandit
    "T",  # flake8-debugger
]
ignore = ["E501", "B008"]

